+++ cni/network/network_linux.go
@@ -1,10 +1,12 @@
 package network
 
 import (
+	"github.com/Azure/azure-container-networking/cni"
+	"github.com/Azure/azure-container-networking/network"
 	cniTypesCurr "github.com/containernetworking/cni/pkg/types/current"
 )
 
 // handleConsecutiveAdd is a dummy function for Linux platform.
-func handleConsecutiveAdd(containerId, endpointId string, nwInfo *NetworkInfo, nwCfg *NetworkConfig) (*cniTypesCurr.Result, error) {
+func handleConsecutiveAdd(containerId, endpointId string, nwInfo *network.NetworkInfo, nwCfg *cni.NetworkConfig) (*cniTypesCurr.Result, error) {
 	return nil, nil
 }
