+++ network/policy/policy.go
@@ -1,0 +1,23 @@
+package policy
+
+import (
+	"encoding/json"
+)
+
+type CNIPolicyType string
+
+type Policy struct {
+	Type CNIPolicyType
+	Data json.RawMessage
+}
+
+// SerializePolicies serializes policies to json.
+func SerializePolicies(policyType CNIPolicyType, policies []Policy) []json.RawMessage {
+	var jsonPolicies []json.RawMessage
+	for _, policy := range policies {
+		if policy.Type == policyType {
+			jsonPolicies = append(jsonPolicies, policy.Data)
+		}
+	}
+	return jsonPolicies
+}
