--- network/policy/policy_linux.go
@@ -1,3 +1,0 @@
-package policy
-
-type Policy struct{}
