+++ network/policy/policy_windows.go
@@ -1,29 +1,7 @@
 package policy
 
-import (
-	"encoding/json"
-)
-
-type CNIPolicyType string
-
 const (
 	NetworkPolicy     CNIPolicyType = "NetworkPolicy"
 	EndpointPolicy    CNIPolicyType = "EndpointPolicy"
 	OutBoundNatPolicy CNIPolicyType = "OutBoundNatPolicy"
 )
-
-type Policy struct {
-	Type CNIPolicyType
-	Data json.RawMessage
-}
-
-// SerializePolicies serializes policies to json.
-func SerializePolicies(policyType CNIPolicyType, policies []Policy) []json.RawMessage {
-	var jsonPolicies []json.RawMessage
-	for _, policy := range policies {
-		if policy.Type == policyType {
-			jsonPolicies = append(jsonPolicies, policy.Data)
-		}
-	}
-	return jsonPolicies
-}
